Add the RX part
The RX part works with the pca10050 board, for which we won't,
however, introduce any BSP.

This commit initializes the RX board with a simgle empty main.c
file and the framework needed to build (no flashing yet, though).

Also, we change the Makefile.posix to reflect the toolchain we use.