Add detection of ``our'' device
Now we use local name to see if this is the device we want to
connect. There may be other criteria to determine our target.
Please see the comment introduced by this patch and keep this
comment up to date if you touch this code.

Expected: application reports to the serial console when it has
found the expected target device presently advertising.