RX: restart advertising upon disconnect
Before this patch the receiver went dead after BT link disconnected.
Now we restart advertising.

While at it, we added a new timer to blink the green LED while
the device in in the advertising state. This should help testing.
To disable this feature, uncomment the

line, and the whole thing will be preprocessed out.