TX TEST: disable gyro and minimise SPI traffic
We need to isolate the reboot-on-disconnect problem, so first
let's see if it is the SPI traffic that gets into the way of
Bluetooth communication.

This patch gets back to the slow timer (1 Hz) and counter, which
we already had a few commits ago: TX goes to disconnect when
the counter reaches 10.

Check if the nRF51 gets restarted on disconnect in this setting.