TX: introduce new state to prevent unneeded disconnect request
We had bogus reboots, because the timer fired _while_
the BT stack was in the midst of disconnection, and
requested another disconnect. From that, the BTLE stack
went crazy, which caused an unnecessary reboot.

Now we don't do anything in the timer handler when the
BTLE stack is busy disconnecting.