TX: introduce the user activity tracking timer, Part I
We want to be sure that power-hungry devices such as radio
and gyroscope are switched off when the user is inactive.

With this patch we introduce another timer, which tracks the
user activity at the TX side. This patch created the timer,
and a few upcoming patches will implement the functionality.